Move Edit Menu into separate file.
